src/qa_app/requirements.txt
streamlit
st-annotated-text
flask